requirements.txt
fastapi==0.115.8  # Or latest
pydantic==2.10.6  # Or latest
langchain==0.3.17
langchain-community==0.3.16
langchain-google-genai==2.0.9 #Or latest
langchain_core==0.3.33
chromadb==0.4.15 # Or latest, but ensure compatibility with langchain
python-dotenv==1.0.0 # Or latest
uvicorn==0.23.2 # Or latest
